Decouple unit and integration tests for python-sdk (#1487)
# Description
closes: https://github.com/astronomer/astro-sdk/issues/657
## What is the current behavior?
We still have some unit and integrations tests exist in the same file

## What is the new behavior?
moves unit tests in test/  and integration test in test_integration

## Does this introduce a breaking change?
No

### Checklist
- [ ] Created tests which fail without the change (if possible)
- [ ] Extended the README / documentation, if necessary
