+++ python-sdk/tests/databricks_tests/test_databricks_api_util.py
@@ -1,6 +1,9 @@
 import pathlib
 from unittest import mock
-from unittest.mock import call
+from unittest.mock import Mock, call
+
+import pytest
+from requests.exceptions import HTTPError
 
 from astro.databricks.api_utils import (
     create_and_run_job,
@@ -10,6 +13,35 @@ from astro.databricks.api_utils import (
 )
 
 CWD = pathlib.Path(__file__).parent
+
+
+def raise_arb_http_error(unneeded_arg):
+    arbitrary_http = HTTPError()
+    arbitrary_http.response = Mock(json=lambda: {"error_code": "FOOBAR"})
+    raise arbitrary_http
+
+
+def raise_not_exist_error(unneeded_arg):
+    non_existent_http = HTTPError()
+    non_existent_http.response = Mock(json=lambda: {"error_code": "RESOURCE_DOES_NOT_EXIST"})
+
+
+@mock.patch("databricks_cli.secrets.api.SecretApi.delete_scope")
+@mock.patch("databricks_cli.sdk.api_client.ApiClient")
+def test_delete_scope_http_error_arbitary(mock_api_client, mock_delete_secret):
+    """Raise error when delete_scope throw error"""
+    mock_delete_secret.raiseError.side_effect = raise_arb_http_error
+    mock_delete_secret.side_effect = raise_arb_http_error
+    with pytest.raises(HTTPError):
+        delete_secret_scope("non-existent-scope", api_client=mock_api_client)
+
+
+@mock.patch("databricks_cli.secrets.api.SecretApi.delete_scope")
+@mock.patch("databricks_cli.sdk.api_client.ApiClient")
+def test_delete_scope_http_error_non_existent(mock_api_client, mock_delete_secret):
+    """Do not raise an error when trying to delete non-existent scope and the error code is RESOURCE_DOES_NOT_EXIST"""
+    mock_delete_secret.side_effect = raise_not_exist_error
+    delete_secret_scope("non-existent-scope", api_client=mock_api_client)
 
 
 @mock.patch("databricks_cli.sdk.api_client.ApiClient")
