+++ python-sdk/tests/sql/operators/test_dataframe.py
@@ -1,3 +1,4 @@
+import os
 import pathlib
 from unittest import mock
 
@@ -5,6 +6,7 @@ import pandas
 import pytest
 from airflow.configuration import conf
 from airflow.exceptions import AirflowException
+from airflow.models.xcom import BaseXCom
 from airflow.utils import timezone
 
 import astro.sql as aql
@@ -210,3 +212,26 @@ def test_col_case_is_preserved(sample_dag):
         task1 = sample_df_1()
         validate(task1)
     test_utils.run_dag(sample_dag)
+
+
+@mock.patch("airflow.models.xcom.XCom", BaseXCom)
+@mock.patch("astro.custom_backend.serializer")
+@mock.patch.dict(os.environ, {"AIRFLOW__CORE__ENABLE_XCOM_PICKLING": "True"})
+def test_dataframe_from_file_xcom_pickling(mock_serde, sample_dag):
+    """Assert when ENABLE_XCOM_PICKLING is true we do not use custom backend serializer"""
+
+    @aql.dataframe
+    def validate_file(df: pandas.DataFrame):  # skipcq: PY-D0003
+        assert len(df) == 5
+        assert "sell" in df.columns
+        return df
+
+    @aql.dataframe()
+    def count_df(df: pandas.DataFrame):
+        return len(df)
+
+    with sample_dag:
+        count_df(validate_file(df=File(path=str(CWD) + "/../../data/homes2.csv")))
+    test_utils.run_dag(sample_dag)
+    mock_serde.serialize.assert_not_called()
+    mock_serde.deserialize.assert_not_called()
