+++ python-sdk/tests/sql/operators/test_export_file.py
@@ -6,11 +6,12 @@ import pytest
 
 import astro.sql as aql
 from astro.airflow.datasets import DATASET_SUPPORT
-from astro.constants import Database
+from astro.constants import Database, FileType
 from astro.files import File
 
 # Import Operator
-from astro.sql.operators.export_table_to_file import export_table_to_file
+from astro.sql import ExportFileOperator, export_file
+from astro.sql.operators.export_table_to_file import ExportTableToFileOperator, export_table_to_file
 from astro.table import Table
 
 from ..operators import utils as test_utils
@@ -145,3 +146,49 @@ def test_inlets_outlets_non_supported_ds():
     )
     assert task.operator.inlets == []
     assert task.operator.outlets == []
+
+
+def test_raise_exception_for_invalid_input_type():
+    """Raise value error when input data is not correct in ExportTableToFileOperator"""
+    with pytest.raises(ValueError) as exc_info:
+        ExportTableToFileOperator(
+            task_id="task_id",
+            input_data=123,
+            output_file=File(
+                path="gs://astro-sdk/workspace/openlineage_export_file.csv",
+                conn_id="bigquery",
+                filetype=FileType.CSV,
+            ),
+            if_exists="replace",
+        ).execute(context=None)
+    expected_msg = "Expected input_table to be Table or dataframe. Got <class 'int'>"
+    assert exc_info.value.args[0] == expected_msg
+
+
+# TODO: Remove this test in astro-sdk 2.0.0
+def test_warnings_message():
+    """Assert the warning log when using deprecated method"""
+    with pytest.warns(
+        expected_warning=DeprecationWarning,
+        match="""This class is deprecated.
+            Please use `astro.sql.operators.export_table_to_file.ExportTableToFileOperator`.
+            And, will be removed in astro-sdk-python>=2.0.0.""",
+    ):
+        ExportFileOperator(
+            task_id="task_id",
+            input_data=123,
+            output_file=File(
+                path="gs://astro-sdk/workspace/openlineage_export_file.csv",
+                conn_id="bigquery",
+                filetype=FileType.CSV,
+            ),
+            if_exists="replace",
+        )
+
+    with pytest.warns(
+        expected_warning=DeprecationWarning,
+        match="""This decorator is deprecated.
+        Please use `astro.sql.operators.export_table_to_file.export_table_to_file`.
+        And, will be removed in astro-sdk-python>=2.0.0.""",
+    ):
+        export_file(input_data=Table(), output_file=File(path="/tmp/saved_df.csv"), if_exists="replace")
