+++ python-sdk/tests_integration/databases/test_bigquery.py
@@ -25,6 +25,7 @@ SUPPORTED_CONN_IDS = [DEFAULT_CONN_ID, CUSTOM_CONN_ID]
 CWD = pathlib.Path(__file__).parent
 
 
+@pytest.mark.integration
 @pytest.mark.parametrize("conn_id", SUPPORTED_CONN_IDS)
 def test_create_database(conn_id):
     """Test creation of database"""
@@ -467,6 +468,7 @@ def test_export_table_to_file_file_already_exists_raises_exception(
     assert err_msg.endswith(f"The file {filepath} already exists.")
 
 
+@pytest.mark.integration
 def test_get_schema_location():
     db = BigqueryDatabase(conn_id="gcp_conn")
     location = db._get_schema_location("tmp_astro")
