+++ python-sdk/tests_integration/sql/operators/test_dataframe.py
@@ -1,10 +1,7 @@
-import os
 import pathlib
-from unittest import mock
 
 import pandas
 import pytest
-from airflow.models.xcom import BaseXCom
 from airflow.utils import timezone
 
 from astro import sql as aql
@@ -335,24 +332,3 @@ def test_dataframe_append_table_if_exist(sample_dag, conn_id):
     with sample_dag:
         aql.drop_table(table=output_tb)
     test_utils.run_dag(sample_dag)
-
-
-@mock.patch("airflow.models.xcom.XCom", BaseXCom)
-@mock.patch("astro.custom_backend.serializer")
-@mock.patch.dict(os.environ, {"AIRFLOW__CORE__ENABLE_XCOM_PICKLING": "True"})
-def test_dataframe_from_file_xcom_pickling(mock_serde, sample_dag):
-    @aql.dataframe
-    def validate_file(df: pandas.DataFrame):  # skipcq: PY-D0003
-        assert len(df) == 5
-        assert "sell" in df.columns
-        return df
-
-    @aql.dataframe()
-    def count_df(df: pandas.DataFrame):
-        return len(df)
-
-    with sample_dag:
-        count_df(validate_file(df=File(path=str(CWD) + "/../../data/homes2.csv")))
-    test_utils.run_dag(sample_dag)
-    mock_serde.serialize.assert_not_called()
-    mock_serde.deserialize.assert_not_called()
