+++ python-sdk/tests_integration/sql/operators/transform/test_postgres_transform.py
@@ -31,6 +31,7 @@ def validate(df: pd.DataFrame):
     assert df.iloc[0].to_dict()["first_name"] == "PENELOPE"
 
 
+@pytest.mark.integration
 @pytest.mark.parametrize(
     "output_table",
     [Table(), Table(name="my_table"), Table(name="my_table", conn_id="postgres_conn_pagila")],
@@ -98,6 +99,7 @@ def pg_query_result(request):
         )
 
 
+@pytest.mark.integration
 @pytest.mark.parametrize(
     "pg_query_result",
     ["basic", "semicolon", "with_param", "with_jinja", "with_jinja_template_params"],
@@ -120,6 +122,7 @@ def test_postgres(sample_dag, pg_query_result):
     test_utils.run_dag(sample_dag)
 
 
+@pytest.mark.integration
 @pytest.mark.parametrize(
     "database_table_fixture",
     [
@@ -158,6 +161,7 @@ def test_postgres_join(sample_dag, database_table_fixture):
         validate(ret)
 
 
+@pytest.mark.integration
 def test_postgres_set_op_kwargs(sample_dag):
     @aql.dataframe
     def validate_result(df: pd.DataFrame):
